Fixes issue 44. Turns out to be a bug in the Http plugin.
The Http plugin decides what format to return the data in based on the content-type of the response, but the regex to recognise an XML response assumes that it will not contain anything else other than mime-type. Last.fm returns the content type "text/xml; charset=utf-8" which meant the Http returned the response as a string, rather than the expected SimpleXMLElement.