Ensure that heartbeat tick always fires.
This change addresses the problem where the server connection has
dropped, and the Ping plugin can't detect it.  This is because
on EOF, getActiveReadSockets() doesn't return the hostmask for
the connection, and therefore the connection's onTick() never fires.

Now, we fire the onTick() and processEvents(), which handles
pending outgoing messages to the server, every iteration, and
only try to read events from the connection if getActiveReadSockets()
indicates there should be something waiting.