Fail php plugin if db is not loaded
When loading the php plugin failed caused by an unreadable db of the plugin,
an error will be throwen by PDO in Phergie_Plugin_Php_Source_Local.

This error is caught, but we do not get an instance of the db connection.
So instead of failing, the plugin continues to run until the plugin needs
the connection, and because the connection not an object, but null, no
method operation can be done on $this->connection in
Phergie_Plugin_Php_Source_Local.