Learn phergie to locate the configuration file on multiple locations
There was an issue where pear installed the configuration on a wrong location.
A way to solve this problem is to force Phergie to actually check if the file
exists before opening that specific configuration file.

With this commit, there are 2 paths to be tested. If these attemps fails,
Phergie will show an error that the user should try
"phergie path/to/Settings.php".

- Fixes #174