+Twitter OAuth Tweeting
Implementation of Tweeting through OAuth.

Configuration
------------

It is required that the user create an application with both read and
write access through Twitter Developer API and authenticate their
account for said application.  This information should then be stored
within the config file loaded with Phergie.

It is possible to hide a consumerkey and consumersecret, but I believe
that most users will have no problem with the current limitation.  Once
more basic functionality has been address, it should be modified so that
users have the option of using their own application, or one provided by
default.

* twitter.consumerkey
* twitter.consumersecret
* twitter.usertoken
* twitter.usersecret

Application
----------

PEAR module/extension OAuth is now required.  I am looking for advice as
to how you would prefer this requirement included.  I feel as though
requiring it when the user has not provided authentication credentials
would not be the best idea.  What would be the best way of making OAuth
a requirement only when specific configuration variables have values?

A new function onCommandTweet(); has been added so that users can !tweet
TWEET MESSAGE and send a tweet.  The user is notified on error, or given
the tweet back from Phergie.

Notes
-----

The four oauth credentials are private variables, but the oauth object
itself is public.  Since the oauth object belongs to $twitter, which is
private, this should not cause any problems with other plugins accessing
the data.

My experience with OOP is very limited.  Please note this when reviewing
the code.  I welcome any criticism and explanation for any of the
methods I used that you may have done differently.

I also fixed an issue where Twitter URLs were not being parsed correctly
for their ids.  Phergie will now respond with the tweet message of a
twitter URL linked within the chat.