Removing 'nick change' logging.
Turns out that you can't log a nick change in a 'room', it just comes
through as a server notification.  So if I logged it, you'd then have
to work in the logging viewer to show it in all channels on that
server.  Not worth it at the moment for this simple plugin.